Búsqueda dentro de la página de anuncios totalmente funcional

+++ Anuncios.php
@@ -7,10 +7,10 @@
     <body>
         <div class="upperbar-container">
             <div class="searchbar" style="float:left;">
-                <form id="form"> 
+                <form method="post"> 
                     <input class="buscar" type="search" id="query" name="q" size="50" placeholder="Buscar artículos...">
                     <button class="buscar">Filtros</button>
-                    <button class="buscar" type="submit">Buscar</button>
+                    <button class="buscar" type="submit" name="boton-buscar">Buscar</button>
                 </form>
             </div>
 
@@ -43,11 +43,15 @@
                     $consulta_articulos = mysqli_query($conexion, "select * from Articulo") or die ("Fallo en la consulta de Articulos");
 
                     if($_SERVER["REQUEST_METHOD"] == "POST"){
-                        $id_articulo_deseos=$_POST["Id_Articulo"];
-                        if($_POST["add-listadeseados"] == 'delete'){ #En este caso significa que queremos eliminar el archivo de la lista de deseos
-                            mysqli_query($conexion, "delete from Deseos where id_Articulo=$id_articulo_deseos and id_Usuario=$logged_user");
-                        }else{ #En este caso significa que queremos añadir el archivo a la lista de deseos
-                            mysqli_query($conexion, "insert into Deseos values($logged_user, $id_articulo_deseos)");
+                        if(isset($_POST["Id_Articulo"])){ #Cuando queremos añadir o eliminar un articulo de la lista de deseos
+                            $id_articulo_deseos=$_POST["Id_Articulo"];
+                            if($_POST["add-listadeseados"] == 'delete'){ #En este caso significa que queremos eliminar el archivo de la lista de deseos
+                                mysqli_query($conexion, "delete from Deseos where id_Articulo=$id_articulo_deseos and id_Usuario=$logged_user");
+                            }else{ #En este caso significa que queremos añadir el archivo a la lista de deseos
+                                mysqli_query($conexion, "insert into Deseos values($logged_user, $id_articulo_deseos)");
+                            }
+                        }elseif(isset($_POST['boton-buscar']) && isset($_POST['q']) && ($_POST['q'] != "")){ #Cuando queremos buscar un artículo
+                            $consulta_articulos = mysqli_query($conexion, 'select * from Articulo where Nombre like "%'.$_POST["q"].'%"') or die ("Fallo en la consulta de Articulos");
                         }
                     }
 
@@ -83,7 +87,7 @@
                         #echo '<input type="hidden" name="Id_Articulo" value="'.$fila['Id_Articulo'].'">';
                         #echo '<input type="hidden" name="Id_Usuario" value="'.$logged_user.'">';
                         echo '</button></div>';
-                        echo '<a href="#"><img class="anuncio" height="300" width="250" src="../ImagenesArticulos/'.$fila['Imagen'].'.jpg"/></a>'; #Insertamos la imagen con el hipervínculo a la página del anuncio
+                        echo '<a href="#"><img class="anuncio" height="300" width="250" src="./ImagenesArticulos/'.$fila['Imagen'].'.jpg"/></a>'; #Insertamos la imagen con el hipervínculo a la página del anuncio
                         echo '<br>';
                         echo '<a class="texto-anuncio" href="#"><div align="center">'.$fila['Nombre'].' - '.$fila['Precio'].'€</div></a>'; #Insetamos el nombre y el precio con el hipervínculo a la página del anuncio
                         echo '</div></td>';
